requirements.txt
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==5.2.6
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mysqlclient==2.2.7
PyJWT==2.10.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
setuptools==80.9.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
